requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.6.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-SQLAlchemy==2.4.0
http-ece==1.1.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
pluggy==0.9.0
psycopg2==2.8.2
py==1.8.0
py-vapid==1.4.0
pycparser==2.19
pymongo==3.7.2
pytest==4.4.0
pywebpush==1.9.2
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.24.1
Werkzeug==0.15.2

# 
hyper
jwt

# 
pyOpenSSL==19.1.0